+++ Mvk/MvkClient/Entity/Particle/EntityTestFX.cs
@@ -12,7 +12,7 @@ namespace MvkClient.Entity.Particle
         {
             particleGravity = .02f;
             textureUV = new vec2i(0, 9);
-            color = new vec3((float)rand.NextDouble(), (float)rand.NextDouble(), (float)rand.NextDouble());
+            color = new vec3(rand.NextFloat(), rand.NextFloat(), rand.NextFloat());
             particleMaxAge *= 2;
         }
